Converted result set into an iterator for less memory wastage.

+++ scripts/benchmark.php
@@ -1,5 +1,7 @@
 <?php
 
+define('NODE_COUNT',100000);
+
 require_once(dirname(__FILE__).'/../lib/octopi.php');
 
 $pdo = new PDO('mysql:host=localhost;dbname=graphtest',
@@ -9,24 +11,32 @@ $pdo = new PDO('mysql:host=localhost;dbname=graphtest',
 );
 
 $graph = new Octopi_Graph($pdo);
-$graph->truncate();
 
-printf("generating 1000000 nodes\n");
-$time = microtime(true);
-
-for($i=1; $i<=1000000; $i++)
+if(in_array('--generate', $argv))
 {
-	if($i % 100 == 0) printf("generating node %s\n", $i);
+	$graph->truncate();
 
-	$node = $graph->createNode(array(), $i);
+	printf("generating %s nodes\n", number_format(NODE_COUNT));
+	$time = microtime(true);
 
-	if($i>1)
+	for($i=1; $i<=NODE_COUNT; $i++)
 	{
-		$graph->node(rand(1,$i-1))->createEdge($node);
-	}
-}
+		if($i % 100 == 0) printf("generating node %s\n", $i);
+		$node = $graph->createNode(array(), $i);
 
-printf("time elapsed: %fs\n", microtime(true)-$time);
+		// pick four random nodes
+		$p1 = $graph->node(rand(1,$i-1));
+		$p2 = $graph->node(rand(1,$i-1));
+
+		if($i>1)
+		{
+			// use the rich-get-richer method to build a natural graph
+			$node->createEdge($p1->degree() > $p2->degree() ? $p1 : $p2);
+		}
+	}
+
+	printf("time elapsed: %fs\n", microtime(true)-$time);
+}
 
 for($i=0; $i<6; $i++)
 {
@@ -36,10 +46,10 @@ for($i=0; $i<6; $i++)
 	$traversal = new Octopi_Traversal(array(
 		'start'=>$graph->n1,
 		'depth'=>$i,
-		'direction'=>Octopi_Edge::OUT
+		'direction'=>Octopi_Edge::EITHER
 		));
 
 	$result = $graph->traverse($traversal);
-	printf("found %d nodes in %fs\n", $result->count(), microtime(true)-$time);
+	printf("found %d paths in %fs\n", $result->count(), microtime(true)-$time);
 	printf("peak memory usage: %dKb\n", memory_get_peak_usage(true) / 1024);
 }
